AUR, Cleanup, and New Options
AUR support is now available.
Works with both Aura and Yaourt automatically.
Packages that are for AUR will be displayed if no helper is available.
su was modified to do checks a bit cleaner.
db was refactored to house the common package query function.
main was refactored to be a bit less verbose. Verbose and debug were
removed and some code to look cleaner.
Refresh and upgrade options were created to choose if one wanted to
refresh or upgrade before installation of new packages.