Add Clippy
Clippy as a service was alright but it's easier when it's integrated
into the code itself as a plugin. Since shaper is using the nightly
compiler for now anyways due to diesel it seemed better to peg clippy at
the version that works with the compiler day I've chosen for the project.
This commit also cleans up any suggestions from the linter. It seems to
have caught mostly things I had figured out but hadn't implemented
everywhere, such as checking if the file exists and using a collapsed
if statement.