V0.1.0 (#12)
* Initial SQL Commit

Creates files to be used for PostgreSQL Database.
tables.sql focuses on creating all needed tables
arch.sql focuses on creating arch specific mappings
sql_setup.sh will create the database for the end user w/ no hassle

* env, SQL, and Diesel

Adds Diesel dependencies
Adds SQL to project
Adds env to project
Changes Cargo manifest to reflect additions
Changes travis to work with the correct nightly
Updates README
Code changes to replace inputs with correct versions on arch

* Schema Change

Adds ubuntu-dev packages to schema

* AUR, Cleanup, and New Options

AUR support is now available.
Works with both Aura and Yaourt automatically.
Packages that are for AUR will be displayed if no helper is available.
su was modified to do checks a bit cleaner.
db was refactored to house the common package query function.
main was refactored to be a bit less verbose. Verbose and debug were
removed and some code to look cleaner.
Refresh and upgrade options were created to choose if one wanted to
refresh or upgrade before installation of new packages.

* Add Clippy

Clippy as a service was alright but it's easier when it's integrated
into the code itself as a plugin. Since shaper is using the nightly
compiler for now anyways due to diesel it seemed better to peg clippy at
the version that works with the compiler day I've chosen for the project.
This commit also cleans up any suggestions from the linter. It seems to
have caught mostly things I had figured out but hadn't implemented
everywhere, such as checking if the file exists and using a collapsed
if statement.

* Alchemist Overhaul

Fixes travis builds that were failing
Changes name from Shaper to Alchemist which is more fitting
Refactors code to be cleaner and better organized
Changes code to reference alchemy as the library not shaper
Adds setup script for Arch Linux
Changes db to alchemist
Updates Cargo dependencies

* Update README.md

Check off last box in v0.1.0