v0.2/v0.3
v0.2.0

Due to the fact that I rewrote everything to use sqlite rather than
using postgresql I'm bumping the version to v0.2.0 rather than doing
a minor bump. The benefit of this though is that many people already
have sqlite installed as a dependency on their system. The other benefit
of it is that it's much lighter on the system in terms of resources
making it real nice for a command line tool like Alchemist. The jump to
v0.2.0 makes sense because everything was switched from Vecs to HashSets
breaking anyone using code based off this.

Release Notes:

- HashSets rather than Vecs for increased efficiency and fitting the needs
  of how Alchemist works
- SQLite rather than Postgresql for lighter dev dependencies and less
  resources being used on the system.

Upgrade

* Error Fixes, Warning Suppression, and Error Handling

Adds better error handling for DB connections when they aren't found
Fixes errors in building the code because of mismatched brackets
Supresses warnings from AUR Enum for unused parts of it
Adds empty helper method to the Package type to create a null version

v0.3.0

Add Void Linux Support
Change how distro checking works
Various administrative tasks and doc updates