Begin rewrite of alchemist client code (#84)
* Begin rewrite of alchemist client code

This is a huge rewrite of alchemist. While it had been split up in
a previous commit in order to make it a client/server based model this
is the first commit that completely changes that for the client code.
It's been reworked to be both more ergonomic from a maintainer's
standpoint and easily extensible for other users. Here are the main
highlights:

- Acts a bit more distro independent by querying for the package
  manager. Later it will be extended to query the distro for the
  package conversion functionality. This makes it easier for cases
  where distros use the same package manager (i.e. apt-get is used for
  Debian and Ubuntu)
- Uses error-chain to handle errors better with Rust. We can chain
  errors, not have to worry about the conversion, and then use ? more
  liberally throughout the code making it more robust and easier to
  maintain.
- macros to reduce boilerplate code have been added to make package
  commands easier to write. Not only is this more maintainable but
  it also makes the trait implementations for package managers are
  more uniform and easier to deal with.

Most of the rewrite focuses on providing a solid base for maintainers
but it also makes it easy for new contributors by making it easy to
read and understand!