Add Alchemist Server
This was a lot of work and theres still things to be done, however it
should be good for now.

Rocket is the backend for the web server with serde for json, it takes
requests at "http://<domain>/api/v0/". There are three endpoints ATM:
    - ``/`` for the index, takes a GET
    - ``/version["/", "/client", "server"]`` returns the current
    versions of the client and server on GitHub respectivly, or both if not
    specified.
    - ``/packages`` Takes POST request with the package[s], distro name,
    the name of the client and version of the client.

This is a pretty big change to how Alchemist will work and is a on going
progress.